+++ packages/misskey-js/src/autogen/types.ts
@@ -4676,6 +4676,8 @@ export type components = {
       title: string;
       summary: string;
       script: string;
+      /** @enum {string} */
+      visibility: 'private' | 'public';
       likedCount: number | null;
       isLiked?: boolean;
     };